+++ CHANGELOG.md
@@ -13,6 +13,7 @@
     * ``Cmd Option E`` - Expand star imports
     * ``Cmd Option R`` - Rename identifier
     * Can be used as scripts as well, see `script/refactoring` folder
+    * Preview dialog can be closed with `Cmd .` / `Esc`, confirmed with `Enter`
 
 ## 1.3.3 (2017-10-27)